Add stub class to indicate location of OSGM02 file
This is purely a hack, but as the orthometric height functions are of limited utility on iOS, this allows us to keep the tests passing whilst also not shipping OSGM02 in the framework bundle. To use OSGM02, switch both the binary file and the `OSGM02Stub` class's inclusion from the test bundle to the framework bundle.